change Module as sigle py file
